+++ CHANGES.md
@@ -3,8 +3,10 @@
 ## version 10.2.07
 - Correction :
     - Bloquage de l'incrémentation de l'id de page lorsque deux pages ont le même nom.
-- Modification :
+    - Login : l'option "Se souvenir de moi" est fonctionnelle.
+- Modifications :
     - Configuration: sauvegarde du site dans une archive : animation d'attente avec message de confirmation ou d'erreur.
+    - Cookies d'authentification : option samesite sur Strict
 
 ## version 10.2.06
 - Corrections :